sdk-aws-python/src/requirements.txt
black==23.7.0
boto3==1.28.20
botocore==1.31.20
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
docker==6.1.3
flake8==6.1.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mccabe==0.7.0
moto==4.1.14
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pip==23.2.1
platformdirs==3.10.0
py-partiql-parser==0.3.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
responses==0.23.3
s3transfer==0.6.1
setuptools==68.0.0
six==1.16.0
tomli==2.0.1
types-PyYAML==6.0.12.11
urllib3==1.26.16
websocket-client==1.6.1
Werkzeug==2.3.6
wheel==0.38.4
xmltodict==0.13.0
